requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.5.7
charset-normalizer==3.2.0
Django==4.2.3
djangorestframework==3.14.0
drf-yasg==1.21.6
idna==3.4
inflection==0.5.1
numpy==1.24.4
packaging==23.1
pandas==2.0.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
